backend/requirements.txt
fastapi
uvicorn
pymongo
python-decouple
python-jose
passlib
